requirements.txt
# Telegram бот
aiogram==3.7.0
redis==5.0.4

# База данных
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Celery
celery==5.3.6
flower==2.0.1

# Извлечение текста
pdfplumber==0.11.0
python-docx==1.1.0

# Шифрование
cryptography==42.0.5

# TTS — раскомментируйте нужный движок

# Вариант 1: CoquiTTS (локально, лучше качество)
TTS==0.22.0

# Вариант 2: ElevenLabs API
# elevenlabs==1.2.0

# Утилиты
python-dotenv==1.0.1
aiofiles==23.2.1
